requirements.txt
langchain
langchain-community
langchain-groq
langchain-google-genai
langchain-ollama
langgraph
llama-index
llama-index-vector-stores-qdrant
llama-index-embeddings-ollama
llama-index-llms-ollama
qdrant-client
python-dotenv
newsapi-python
requests
crewai
litellm
flask
langfuse>=3.0,<4.0
